Refactor imports and remove unused routers from Trace and LocationInTrace APIs
